[OT-347] [REFACTOR]: formatTime 파일 분리
