+++ src/shared/lib/index.ts
@@ -2,3 +2,4 @@ export * from "./cn";
 export * from "./queryClient";
 export { default as QueryProvider } from "./QueryClientProvider";
 export * from "./formatDate";
+export * from "./formatTime";
